+++ packages/common/CHANGELOG.md
@@ -1,8 +1,19 @@
 # Changelog
 
-## Unreleased
+## 7.0.0 2023-05-15 
+
+### Breaking changes
+
+- Updated `blockInfo` so that the `bakerId` field is optional, since it will be undefined for genesis blocks.
+- `waitForTransactionFinalization` now returns a `BlockItemSummaryInBlock`
+- Added missing version return type in `getModuleSchema`. It now returns an object containing the schema source and version.
 
 ### Added
+
+- Helpers for calculating energy cost for a transaction and microCCD cost from energy cost:
+  - `getEnergyCost`
+  - `getExchangeRate`
+  - `convertEnergyToMicroCcd`
 - Utility functions for extracting information from `BlockItemSummary`:
     - `isInitContractSummary`
     - `isUpdateContractSummary`
@@ -20,9 +31,11 @@
     - `findEarliestFinalized`
     - `findInstanceCreation`
     - `findFirstFinalizedBlockNoLaterThan`
+- Extended HdWallet with support for verifiable credential key deriviation.
 
-### Breaking changes
-- `waitForTransactionFinalization` now returns a `BlockItemSummary`
+### Changed
+
+- Bumped @concordium/rust-bindings to 0.12.0. (Adds key derivation for verifiable credentials)
 
 ## 6.5.0 2023-5-03